Finalizada

+++ CalculadoraComSwitch/Calculadora.cs
@@ -13,22 +13,22 @@ namespace Calculadora
         /// <param name="n1">Número 1</param>
         /// <param name="n2">Numéro 2</param>
         /// <returns>Retorna a soma de dois números</returns>
-        public static int Somar(int n1, int n2)
+        public static double Somar(double n1, double n2)
         {
-            return n1 + n2;
+            return  n1 + n2;
         }
-        public static int Subtrai(int n1, int n2)
+        public static double Subtrai(double n1, double n2)
         {
             return n1 - n2;
         }
 
-        public static int Dividir(int n1, int n2)
+        public static double Dividir(double n1, double n2)
         {
             return n1 / n2;
         }
-        public static int Multiplicar(int n1, int n2)
+        public static double Multiplicar(double n1, double n2)
         {
-            return n1 * n2;
+            return n1* n2;
         }
     }
 }
